requirements.txt
# Test runner
pytest>=8.0.0

# Mobile
Appium-Python-Client>=4.0.0

# Web
playwright>=1.40.0

# Config
PyYAML>=6.0.1

# API channel (--channel api)
requests>=2.31.0

# Reporting (optional)
allure-pytest>=2.13.0
